changed BigDecimal to Long

+++ src/main/java/com/asg/shipchandling/salesquotationsch/dto/request/UpdateSalesQuotationSchRequest.java
@@ -20,7 +20,7 @@ public class UpdateSalesQuotationSchRequest {
 
     private Timestamp transactionDate;
 
-    private BigDecimal customerPoid;
+    private Long customerPoid;
 
     private Long addressPoid;
